pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ai-ml-models"
version = "0.1.0"
description = "AI Model Wrappers for OpenAI, Gemini, Claude, Pollinations, and more."
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "."}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.0.0"
google-generativeai = "^0.3.0"
anthropic = "^0.7.0"
huggingface-hub = "^0.19.0"
pillow = "^10.0.0"
httpx = "^0.25.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
flake8 = "^7.0.0"
